Food Collection
Food -
Food is now possible for the snake to collect. Doing so will load the next stage. More flare might be added in the future to indicate this.

2nd Stage-
Temporary stage added to visually show the switching of stages. It is not possible to complete this stage without going out of bounds as there is no third stage yet for the game to load.

Developer Mode-
Added a simple developer mode for the game. This will be used to allow debug abilities and commands. For now, it just allows the snake to fly upward freely, without having to climb.